feat: implemented gamification features and summary
